added all the functions to one file
